renamed intan acquisition file
